faltan las ejcalas, caro es una crack, awante nuestro tp

+++ src/main/java/clases/itinerario.java
@@ -6,7 +6,7 @@ public abstract class itinerario {
     String destinoFinal;
     pasajero pasajero;
 
-    public abstract int tarifaTotal();
+    public abstract double tarifaTotal();
 
     public int getNumItinerario() {
         return numItinerario;
